[Messaging] Update message size on link creation (#44328)
The focus of these changes is to invalidate the
cached maximum message size when the link is created,
and populate with an updated value from the new link.

This is necessary for large message support, where
the maximum message size for entities can be
reconfigred adjusted on the fly.  Because the client
had cached the value, it would not be aware of the
change and would enforce the wrong size for batch
creation.